substitutes operadriver (64 bit) binary for chromedriver (bravedriver download)
- Brave Browser is installed in the "Program Files (x86)" folder and such should be 32-bit, but Brave Browser does not run with the 32-bit chromedriver provided by Google, even though Chrome is supposed to be 64-bit but is also installed in the "Program Files (x86)" folder
  - Chrome somehow works with the win32 chromedriver, but Brave Browser does not work with the win32 chromedriver
  - the exact message (in a python popup window) is:
Unsupported 16-Bit Application
The program or feature ??C:Windowsbravedriver.exe cannnot start or run due to incompatibility with 64-bit versions of Windows. Please contact the software vendor to ask if a 64-bit Windows compatible version is available
- --> so is this implying that the chromedriver uses 16-bit? confusing since Chrome is supposedly 64-bit now but the 32-bit chromedriver on Windows is still compatible

  - the exact message from the stack Traceback is:
OSError: [WinError 216] This version of %1 is not compatible with the version of Windows you're running. Check your computer's system information and then contact the software publisher